+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added a timed cache for the `IPluginView::canResize()` VST3 function, so value
+  will be remembered during an active resize. This makes resizing VST3 plugin
+  editor windows more responsive.
+
 ## [3.2.0] - 2021-05-03
 
 ### Added